Budget dao return type updated to long and Repository too

+++ app/src/main/java/expensetracker/iit/com/expensetracker/Dao/BudgetDao.java
@@ -28,7 +28,7 @@ public interface BudgetDao
     void insertAll(Budget... budgets);
 
     @Insert
-    int insert(Budget budget);
+    long insert(Budget budget);
 
     @Update
     void update(Budget budget);
